Packets decoder now returns decoded packets as simple method results (was using callbacks).

+++ build.sh
@@ -5,7 +5,10 @@
 # Abort if error
 set -xe
 
-rm -r ./dist
+if [ -e "./dist" ]; then
+  rm -r ./dist
+fi
+
 python -m pip install --upgrade build
 python -m build
